Add securitiesInfo and tradingHistory CRUD controllers,templates, validation

+++ securities/src/main/java/ru/megalomaniac/securities/repository/TradingHistoryRepository.java
@@ -3,6 +3,6 @@ package ru.megalomaniac.securities.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.megalomaniac.securities.model.TradingHistory;
 
-public interface TradingHistoryRepository extends JpaRepository<TradingHistory,Long> {
+public interface TradingHistoryRepository extends JpaRepository<TradingHistory,Integer> {
 
 }
